+++ Runtime/Scripts/Input/UIToWorldInputHandler.cs
@@ -7,19 +7,26 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+using UnityEngine.Events;
 using UnityEngine.EventSystems;
 using static PolytopeSolutions.Toolset.GlobalTools.Generic.ObjectHelpers;
 
 namespace PolytopeSolutions.Toolset.Input {
-    public abstract class UIToWorldInputHandler : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler {
+    public class UIToWorldInputHandler : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler {
+        [SerializeField] private UnityEvent onInteractionStarted;
+        [SerializeField] private UnityEvent<RaycastHit> onInteractionPerformed;
+        [SerializeField] private UnityEvent onInteractionEnded;
         [SerializeField]private UIToWorldInputReceiver inputReceiver;
-        
+
         protected void Start() {
             if (!this.inputReceiver) {
                 this.inputReceiver = GameObject.FindObjectOfType<UIToWorldInputReceiver>();
                 if (!this.inputReceiver)
                     this.LogWarning("No UIToWorldInputReceiver found!");
             }
+        }
+        protected void OnDestroy() {
+            this.inputReceiver?.UnregisterInputHandler(this);
         }
 
         ///////////////////////////////////////////////////////////////////////
@@ -28,21 +35,36 @@ namespace PolytopeSolutions.Toolset.Input {
             #if DEBUG2
             this.Log(Pointer Enter");
             #endif
-            this.inputReceiver?.HandlerHoverEnter(this);
+            this.inputReceiver?.RegisterInputHandler(this);
         }
         // Inform the input receiver that this handler is not being hovered over anymore
         public void OnPointerExit(PointerEventData eventData) {
             #if DEBUG2
             this.Log(Pointer Exit");
             #endif
-            this.inputReceiver?.HandlerHoverExit(this);
+            this.inputReceiver?.UnregisterInputHandler(this);
         }
 
         ///////////////////////////////////////////////////////////////////////
         // Methods to be called by the input receiver,
         // if at the moment of press, this handler is being hovered over
-        public abstract void OnInteractionStarted();
-        public abstract void OnInteractionPerformed(Vector2 screenMousePos);
-        public abstract void OnInteractionEnded();
+        public virtual void OnInteractionStarted() { 
+            #if DEBUG2
+            this.Log($"Interaction started");
+            #endif
+            this.onInteractionStarted?.Invoke();
+        }
+        public virtual void OnInteractionPerformed(RaycastHit hitInfo) { 
+            #if DEBUG2
+            this.Log($"Ray hit: {hitinfo.transform.gameObject.name}");
+            #endif
+            this.onInteractionPerformed?.Invoke(hitInfo);
+        }
+        public virtual void OnInteractionEnded() {
+            #if DEBUG2
+            this.Log($"Interaction ended");
+            #endif
+            this.onInteractionEnded?.Invoke();
+        }
     }
 }